Add test script to fuzz the kernel implementations of Adiantum
Add a test script that randomly generates messages, keys, and tweaks,
and verifies that the kernel implementation of Adiantum, as accessed
through AF_ALG, produces the same results as the Python reference
implementation.  Both Adiantum-XChaCha12-AES and Adiantum-XChaCha20-AES
are tested.  An option is also provided to test all known
implementations in the kernel in addition to the default one.