benchmark: avoid undefined behavior in rotate macros
Without this, rotations by 0 bits (such as those that can be done by the
RC5 code) trigger undefined behavior, and thus UBSAN warnings.

Note: it is not necessary to mask the first shift, since it is not
necessary to support rotation amounts that are too large.  The supported
rotation amounts are 0 through 'width - 1'.