benchmark: do not clobber callee-saved NEON registers
Make the ARM assembly code save and restore registers q4-q7 when they
are used, since these are callee-saved registers.

This is not needed in the Linux kernel, or in other contexts where FPU
and NEON code generation is disabled.  IIRC, I had used -mfloat-abi=soft
when testing this code originally.  But that option is not actually
being set by the build system, and building with that option is not
always possible anyway since some systems don't have the softfp library.

I'm now seeing a test failure due to this, so let's fix it.